requirements.txt
pydantic>=2.0.0
openai>=1.0.0
google-genai>=0.1.0
PyPDF2>=3.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
streamlit>=1.30.0
pymupdf>=1.23.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
